requirements.txt
flask
torch
torchvision
timm
scikit-learn
pillow
gunicorn